+++ changelog.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Spatial map implementations now allow you to specify a custom point hashing algorithm to use
+- Added similar hashing algorithm parameter to `GameFramework.Map`
 
 ### Fixed
 - FleeMaps now properly support goal maps where the base map has open edges (fixes #211)
@@ -18,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Applied performance optimizations to GoalMap and FleeMap algorithms
     - ~50% improvement on a full-update operation, using a wide-open (obstacle free) base map
     - Performance gain varies on other test cases but should generally be measurable
+- Defaulted to a usually faster hashing algorithm (one based on the Map's width) in Map's SpatialMaps
 
 ## [3.0.0-alpha08] - 2021-10-17